requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra=dev --no-emit-index-url --no-emit-trusted-host --output-file=requirements-dev.txt pyproject.toml
beartype==0.21.0
    # via tubular (pyproject.toml)
cfgv==3.4.0
    # via pre-commit
coverage==7.9.1
    # via pytest-cov
distlib==0.3.9
    # via virtualenv
exceptiongroup==1.3.0
    # via pytest
filelock==3.18.0
    # via virtualenv
identify==2.6.12
    # via pre-commit
iniconfig==2.1.0
    # via pytest
joblib==1.5.1
    # via scikit-learn
narwhals==1.44.0
    # via tubular (pyproject.toml)
nodeenv==1.9.1
    # via pre-commit
numpy==2.0.2
    # via
    #   pandas
    #   scikit-learn
    #   scipy
packaging==25.0
    # via pytest
pandas==2.3.0
    # via tubular (pyproject.toml)
platformdirs==4.3.8
    # via virtualenv
pluggy==1.6.0
    # via pytest
polars==1.31.0
    # via tubular (pyproject.toml)
pre-commit==4.2.0
    # via tubular (pyproject.toml)
pyarrow==20.0.0
    # via tubular (pyproject.toml)
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via
    #   tubular (pyproject.toml)
    #   pytest-cov
    #   pytest-mock
    #   test-aide
pytest-cov==2.10.1
    # via tubular (pyproject.toml)
pytest-mock==3.14.1
    # via
    #   tubular (pyproject.toml)
    #   test-aide
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via pre-commit
ruff==0.2.2
    # via tubular (pyproject.toml)
scikit-learn==1.6.1
    # via tubular (pyproject.toml)
scipy==1.13.1
    # via scikit-learn
six==1.17.0
    # via python-dateutil
test-aide==0.1.1
    # via tubular (pyproject.toml)
threadpoolctl==3.6.0
    # via scikit-learn
tomli==2.2.1
    # via pytest
typing-extensions==4.14.0
    # via
    #   tubular (pyproject.toml)
    #   exceptiongroup
tzdata==2025.2
    # via pandas
virtualenv==20.31.2
    # via pre-commit
